refactor(web): extract shared sparkline path helpers
Pull the duplicated pathFor/areaFor SVG geometry out of BigSpark and MiniSpark into a colocated sparkPath helper with explicit max and pad parameters.
